adopt omniqueue, remove generic-queue
Replaces the initial (sort of proof-of-concept) `generic-queue` crate
with `omniqueue`.

Since omniqueue offers an _opaque_ representation of the producers and
consumers via the `make_dynaic()` builder option, further refactors can
be made to de-bulk the consumer code.

Currently there's a distinct type for each backend, but at this point we
should be able to reduce this to a single type.
This is planned as a follow-up PR (time allowing).

Since this diff will be an upheaval to the dependency graph no matter
what, I've included a `cargo update` to help keep things current.