WIP: debulk bridge consumer code
Essentially, once we've got omniqueue in bridge, we can "debulk" a lot
of the consumer code by un-traiting it.

A single concrete type can hold an input config which can be used to
produce a `DynConsumer` which can then be used to run the consumer loop.

It's not a huge lift, but I sunk too much time already so I'm parking
the incomplete effort here so I can loop back to it later :(